eval/requirements.txt
accelerate==1.5.2
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
albucore==0.0.17
albumentations==1.4.18
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
astor==0.8.1
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
audioread==3.0.1
av==15.1.0
azure-common==1.1.28
azure-core==1.32.0
azure-identity==1.21.0
azure-mgmt-compute==34.1.0
azure-mgmt-core==1.5.0
azure-mgmt-network==28.1.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.3
blinker==1.9.0
borb==2.1.25
boto3==1.37.20
botocore==1.37.20
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
cli_exit_tools==1.2.7
click==8.1.8
clip==0.2.0
cloudpickle==3.1.1
colorlog==6.10.1
contourpy==1.3.0
crcmod==1.7
cryptography==44.0.2
cssselect==1.3.0
cycler==0.12.1
Cython==3.0.12
dashscope==1.22.2
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
dill==0.3.9
distro==1.9.0
docker==7.1.0
docker-pycreds==0.4.0
dotenv==0.9.9
easyocr==1.7.2
et_xmlfile==2.0.0
eval_type_backport==0.2.2
evdev==1.9.1
exceptiongroup==1.2.2
executing==2.2.1
fabric==3.2.2
Farama-Notifications==0.0.4
fastapi==0.115.12
fastdtw==0.3.4
filelock==3.18.0
fire==0.7.0
Flask==3.0.3
fonttools==4.56.0
formulas==1.2.8
frozenlist==1.5.0
fsspec==2025.3.0
func_timeout==4.3.5
gdown==5.2.0
gitdb==4.0.12
gitignore_parser==0.1.13
GitPython==3.1.44
greenlet==3.1.1
groq==0.20.0
grpcio==1.71.0
grpcio-status==1.71.0
gui-agents==0.2.1
gymnasium==0.28.1
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.29.3
idna==3.10
ImageHash==4.3.2
imageio==2.37.0
importlib_metadata==8.6.1
importlib_resources==6.5.2
invoke==2.2.0
ipdb==0.13.13
ipython==8.18.1
isodate==0.7.2
itsdangerous==2.2.0
jax-jumpy==1.0.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.9.0
jmespath==0.10.0
joblib==1.4.2
json_repair==0.44.1
json5==0.12.1
jsonlines==4.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.7
lazy_loader==0.4
lib-detect-testenv==2.0.8
librosa==0.11.0
llvmlite==0.43.0
lmdb==1.6.2
loguru==0.7.3
lxml==5.3.1
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.7.5
matplotlib-inline==0.1.7
mdurl==0.1.2
modelscope==1.24.0
more-itertools==10.8.0
MouseInfo==0.1.3
mpmath==1.3.0
msal==1.32.0
msal-extensions==1.3.1
msgpack==1.1.0
msgpack-numpy==0.4.8
multidict==6.2.0
multiprocess==0.70.17
mutagen==1.47.0
netifaces==0.11.0
networkx==3.2.1
ninja==1.11.1.4
numba==0.60.0
numpy==1.26.4
numpy-financial==1.0.0
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauth2client==4.1.3
oauthlib==3.3.1
odfpy==1.4.1
openai==1.68.2
opencv-contrib-python==4.11.0.86
opencv-python==4.8.1.78
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
opt-einsum==3.3.0
outcome==1.3.0.post0
packaging==24.2
paddleocr==2.10.0
paddlepaddle==3.0.0
pandas==2.0.3
paramiko==3.5.1
parso==0.8.5
patsy==1.0.2
pdfminer.six==20231228
pdfplumber==0.11.5
pexpect==4.9.0
pillow==11.1.0
pip==25.0
platformdirs==4.3.7
playwright==1.51.0
pluggy==1.6.0
pooch==1.8.2
prompt_toolkit==3.0.52
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psutil==5.9.8
ptyprocess==0.7.0
pure_eval==0.2.3
py==1.11.0
pyacoustid==1.3.0
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAutoGUI==0.9.54
pyclipper==1.3.0.post6
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.10.6
pydantic_core==2.27.2
PyDrive==1.3.1
pyee==12.1.1
pygame==2.6.1
PyGetWindow==0.0.9
Pygments==2.19.1
PyJWT==2.10.1
PyMsgBox==1.0.9
PyMuPDF==1.25.4
PyNaCl==1.5.0
pynput==1.7.8
pyparsing==3.2.3
pypdf==5.4.0
PyPDF2==3.0.1
pypdfium2==4.30.1
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
pytesseract==0.3.13
python-barcode==0.15.1
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.1
python-pptx==1.0.2
python-xlib==0.33
python3-xlib==0.15
pytweening==1.2.0
pytz==2025.2
PyWavelets==1.6.0
PyYAML==6.0.2
qrcode==8.0
qwen-agent==0.0.29
qwen-vl-utils==0.0.14
RapidFuzz==3.12.2
referencing==0.36.2
regex==2024.11.6
requests==2.31.0
requests-file==2.1.0
requests-toolbelt==1.0.0
retry==0.9.2
rich==13.9.4
rpds-py==0.27.1
rsa==4.9
ruamel.yaml==0.19.1
s3transfer==0.11.4
safetensors==0.5.3
schedula==1.5.61
scikit-image==0.24.0
scikit-learn==1.6.1
scipy==1.13.1
selenium==4.30.0
sentry-sdk==2.24.1
setproctitle==1.3.5
setuptools==75.8.0
shapely==2.0.7
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.13.1
soupsieve==2.6
soxr==0.5.0.post1
squirrel-core==0.20.2
stack-data==0.6.3
starlette==0.46.1
statsmodels==0.14.5
sympy==1.13.1
tabulate==0.9.0
termcolor==2.5.0
threadpoolctl==3.6.0
tifffile==2024.8.30
tiktoken==0.9.0
tldextract==5.3.0
together==1.4.6
tokenizers==0.15.2
toml==0.10.2
tomli==2.2.1
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.35.2
trio==0.29.0
trio-websocket==0.12.2
triton==3.1.0
typer==0.15.2
typing_extensions==4.13.0
tzdata==2025.2
uritemplate==4.1.1
urllib3==1.26.20
uvicorn==0.34.0
wandb==0.19.8
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
wrapt_timeout_decorator==1.5.1
wsproto==1.2.0
XlsxWriter==3.2.2
xmltodict==0.14.2
yarl==1.18.3
zipp==3.21.0
